<dgmr> rename 'mfai.pytorch.models.gan_dgmr' to 'mfai.pytorch.models.dgmr'
