+++ mfai/pytorch/lightning_modules/gan_dgmr.py
@@ -19,12 +19,12 @@ import torch
 from lightning import LightningModule
 from torch import Tensor
 
-from mfai.pytorch.losses.gan_dgmr import (
+from mfai.pytorch.losses.dgmr import (
     GridCellLoss,
     loss_hinge_disc,
     loss_hinge_gen,
 )
-from mfai.pytorch.models.gan_dgmr import (
+from mfai.pytorch.models.dgmr import (
     ContextConditioningStack,
     Discriminator,
     Generator,
