+++ tests/test_lightning.py
@@ -12,7 +12,7 @@ from torch import Tensor
 from mfai.pytorch.dummy_dataset import DummyDataModule
 from mfai.pytorch.lightning_modules import SegmentationLightningModule
 from mfai.pytorch.lightning_modules.gan_dgmr import DGMRLightningModule
-from mfai.pytorch.models.gan_dgmr import Discriminator, Generator
+from mfai.pytorch.models.dgmr import Discriminator, Generator
 from mfai.pytorch.models.unet import UNet
 from mfai.pytorch.namedtensor import NamedTensor
